refactor Home.tsx to subcomponents
