+++ src/app/components/sections/BlogList.tsx
@@ -1,0 +1,3 @@
+export function BlogList() {
+  return;
+}
